:sparkles: Add randomized benchmarking generation and state tomography routine
The commit implements a randomized benchmarking circuit generating scheme, given a number of qubits, number of layers, as well as a seed number to reproduce the same examples, if needed. The circuit is then transpiled to a native gate set, and written on a .qasm file.

The commit also includes a state tomography implementation, that has a 4^N scaling for the number of different observables. The expectation values for individual operators, are then used to reconstruct the density matrix.